Work on able to set properties
